requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.5
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.108.0
fonttools==4.49.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-experimental==0.0.53
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.19
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.8.3
mdurl==0.1.2
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
streamlit==1.31.1
tenacity==8.2.3
tesseract==0.1.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
zipp==3.17.0
